# services/learn-service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for managing hiragana and katakana characters",
        "title": "JapaneseStudent Characters API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "shelyahin.mihail@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v3",
    "paths": {
        "/admin/characters": {
            "get": {
                "description": "Get full list of characters ordered by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get list of characters",
                "responses": {
                    "200": {
                        "description": "List of characters",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CharacterListItem"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new character",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Create a character",
                "parameters": [
                    {
                        "description": "Character creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreateCharacterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Character created successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body or character already exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/admin/characters/{id}": {
            "get": {
                "description": "Get full information about a character by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Get character by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Character ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Character information",
                        "schema": {
                            "$ref": "#/definitions/models.Character"
                        }
                    },
                    "400": {
                        "description": "Invalid character ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Character not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a character by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Delete a character",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Character ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid character ID",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Character not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "patch": {
                "description": "Update character fields (partial update)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "Update a character",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Character ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update request",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/models.UpdateCharacterRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Character not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
        "/characters": {
            "get": {
                "description": "Get a list of all hiragana or katakana characters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "characters"
                ],
                "summary": "Get all characters",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Alphabet type: hr (hiragana) or kt (katakana), default: hr",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Locale: en (English), ru (Russian), or de (German - treated as English), default: en",
                        "name": "locale",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of characters",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CharacterResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error - failed to retrieve characters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/characters/row-column": {
            "get": {
                "description": "Get a row or column of hiragana or katakana characters filtered by consonant or vowel",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "characters"
                ],
                "summary": "Get characters by row or column",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Alphabet type: hr (hiragana) or kt (katakana), default: hr",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Locale: en (English), ru (Russian), or de (German - treated as English), default: en",
                        "name": "locale",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Consonant or vowel character",
                        "name": "character",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of characters matching the filter",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CharacterResponse"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - character parameter is required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error - failed to retrieve characters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/characters/{id}": {
            "get": {
                "description": "Get a hiragana or katakana character by its ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "characters"
                ],
                "summary": "Get character by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Character ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Locale: en (English), ru (Russian), or de (German - treated as English), default: en",
                        "name": "locale",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Character details",
                        "schema": {
                            "$ref": "#/definitions/models.Character"
                        }
                    },
                    "400": {
                        "description": "Bad request - id parameter is required or invalid id parameter",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Not found - character not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error - failed to retrieve character",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/tests/{type}/reading": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Get random characters for reading test. Requires authentication.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "tests"
                ],
                "summary": "Get reading test",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Alphabet type: hiragana or katakana",
                        "name": "type",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Locale: en (English), ru (Russian), or de (German - treated as English), default: en",
                        "name": "locale",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of characters to return, default: 10",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of random characters for reading test",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ReadingTestItem"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request - type parameter is required or invalid alphabet type/locale/count",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - authentication required or invalid/expired token",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error - failed to retrieve test characters",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Character": {
            "type": "object",
            "properties": {
                "consonant": {
                    "description": "defines a membership to a consonant group",
                    "type": "string"
                },
                "englishReading": {
                    "type": "string"
                },
                "hiragana": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "katakana": {
                    "type": "string"
                },
                "russianReading": {
                    "type": "string"
                },
                "vowel": {
                    "description": "defines a membership to a vowel group",
                    "type": "string"
                }
            }
        },
        "models.CharacterListItem": {
            "type": "object",
            "properties": {
                "consonant": {
                    "type": "string"
                },
                "hiragana": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "katakana": {
                    "type": "string"
                },
                "vowel": {
                    "type": "string"
                }
            }
        },
...
        "models.CreateCharacterRequest": {
            "type": "object",
            "properties": {
                "consonant": {
                    "type": "string"
                },
                "englishReading": {
                    "type": "string"
                },
                "hiragana": {
                    "type": "string"
                },
                "katakana": {
                    "type": "string"
                },
                "russianReading": {
                    "type": "string"
                },
                "vowel": {
                    "type": "string"
                }
            }
        },
...
        "models.UpdateCharacterRequest": {
            "type": "object",
            "properties": {
                "consonant": {
                    "type": "string"
                },
                "englishReading": {
                    "type": "string"
                },
                "hiragana": {
                    "type": "string"
                },
                "katakana": {
                    "type": "string"
                },
                "russianReading": {
                    "type": "string"
                },
                "vowel": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}